requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
python-dotenv==1.0.1
gunicorn==22.0.0
cryptography>=42.0.0
